requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
plotly>=5.20.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
anthropic>=0.25.0
numpy>=1.26.0
